Header file rationalisation - move unnecessary definitions out of public API
