+++ src/c/device.h
@@ -10,6 +10,7 @@
 #define _EDGEX_DEVICE_DEVICE_H_ 1
 
 #include "edgex/devsdk.h"
+#include "rest_server.h"
 
 extern int edgex_device_handler_device
 (
